fix(andorid): fix error when clicking on a point annotation's callout
