+++ android/src/main/java/com/rnmapbox/rnmbx/components/mapview/RNMBXMapView.kt
@@ -34,18 +34,10 @@ import com.mapbox.maps.extension.style.layers.properties.generated.ProjectionNam
 import com.mapbox.maps.extension.style.layers.properties.generated.Visibility
 import com.mapbox.maps.extension.style.projection.generated.Projection
 import com.mapbox.maps.extension.style.projection.generated.setProjection
-import com.mapbox.maps.plugin.annotation.Annotation
-import com.mapbox.maps.plugin.annotation.AnnotationConfig
-import com.mapbox.maps.plugin.annotation.annotations
-import com.mapbox.maps.plugin.annotation.generated.*
 import com.mapbox.maps.plugin.attribution.attribution
 import com.mapbox.maps.plugin.compass.compass
 import com.mapbox.maps.plugin.delegates.listeners.*
 import com.mapbox.maps.plugin.gestures.*
-import com.mapbox.maps.plugin.locationcomponent.DefaultLocationProvider
-import com.mapbox.maps.plugin.locationcomponent.LocationConsumer
-import com.mapbox.maps.plugin.locationcomponent.LocationProvider
-import com.mapbox.maps.plugin.locationcomponent.location
 import com.mapbox.maps.plugin.logo.logo
 import com.mapbox.maps.plugin.scalebar.scalebar
 import com.mapbox.maps.viewannotation.ViewAnnotationManager
@@ -72,11 +64,9 @@ import com.rnmapbox.rnmbx.events.MapClickEvent
 import com.rnmapbox.rnmbx.events.constants.EventTypes
 import com.rnmapbox.rnmbx.utils.*
 import com.rnmapbox.rnmbx.utils.extensions.toReadableArray
-import com.rnmapbox.rnmbx.v11compat.annotation.AnnotationID
-import com.rnmapbox.rnmbx.v11compat.annotation.INVALID_ANNOTATION_ID
 import java.util.*
 
-import com.mapbox.maps.MapboxMap.*;
+import com.rnmapbox.rnmbx.components.annotation.RNMBXPointAnnotationCoordinator
 import com.rnmapbox.rnmbx.components.images.ImageManager
 
 import com.rnmapbox.rnmbx.v11compat.event.*
@@ -85,7 +75,6 @@ import com.rnmapbox.rnmbx.v11compat.mapboxmap.*
 import com.rnmapbox.rnmbx.v11compat.ornamentsettings.*
 import org.json.JSONException
 import org.json.JSONObject
-import java.util.*
 
 fun <T> MutableList<T>.removeIf21(predicate: (T) -> Boolean): Boolean {
     var removed = false
@@ -225,12 +214,12 @@ open class RNMBXMapView(private val mContext: Context, var mManager: RNMBXMapVie
 
     private val mSources: MutableMap<String, RNMBXSource<*>>
     private val mImages: MutableList<RNMBXImages>
-    public val pointAnnotationManager: RNMBXPointAnnotationManager by lazy {
+    public val pointAnnotations: RNMBXPointAnnotationCoordinator by lazy {
         val gesturesPlugin: GesturesPlugin = mapView.gestures
         gesturesPlugin.removeOnMapClickListener(this)
         gesturesPlugin.removeOnMapLongClickListener(this)
 
-        val result = RNMBXPointAnnotationManager(mapView)
+        val result = RNMBXPointAnnotationCoordinator(mapView)
 
         gesturesPlugin.addOnMapClickListener(this)
         gesturesPlugin.addOnMapLongClickListener(this)
@@ -445,7 +434,7 @@ open class RNMBXMapView(private val mContext: Context, var mManager: RNMBXMapVie
             feature = childView
         } else if (childView is RNMBXPointAnnotation) {
             val annotation = childView
-            pointAnnotationManager.add(annotation)
+            pointAnnotations.add(annotation)
             feature = childView
         } else if (childView is RNMBXMarkerView) {
             feature = childView
@@ -481,7 +470,7 @@ open class RNMBXMapView(private val mContext: Context, var mManager: RNMBXMapVie
             mSources.remove(feature.iD)
         } else if (feature is RNMBXPointAnnotation) {
             val annotation = feature
-            pointAnnotationManager.remove(annotation)
+            pointAnnotations.remove(annotation)
         } else if (feature is RNMBXImages) {
             mImages.remove(feature)
         }
@@ -729,13 +718,15 @@ open class RNMBXMapView(private val mContext: Context, var mManager: RNMBXMapVie
                     )
             ) { features ->
                 if (features.isValue) {
-                    if (features.value!!.size > 0) {
-                        val featuresList = ArrayList<Feature?>()
-                        for (i in features.value!!) {
-                            featuresList.add(i.feature)
+                    features.value?.let { features ->
+                        if (features.size > 0) {
+                            val featuresList = ArrayList<Feature?>()
+                            for (i in features) {
+                                featuresList.add(i.feature)
+                            }
+                            hits[source.iD] = featuresList
+                            hitTouchableSources.add(source)
                         }
-                        hits[source.iD] = featuresList
-                        hitTouchableSources.add(source)
                     }
                 } else {
                     Logger.e("handleTapInSources", features.error ?: "n/a")
@@ -747,11 +738,11 @@ open class RNMBXMapView(private val mContext: Context, var mManager: RNMBXMapVie
 
     override fun onMapClick(point: Point): Boolean {
         val _this = this
-        if (pointAnnotationManager.getAndClearAnnotationClicked()) {
+        if (pointAnnotations.getAndClearAnnotationClicked()) {
             return true
         }
         if (deselectAnnotationOnTap) {
-            if (pointAnnotationManager.deselectSelectedAnnotation()) {
+            if (pointAnnotations.deselectSelectedAnnotation()) {
                 return true
             }
         }
@@ -783,7 +774,7 @@ open class RNMBXMapView(private val mContext: Context, var mManager: RNMBXMapVie
 
     override fun onMapLongClick(point: Point): Boolean {
         val _this = this
-        if (pointAnnotationManager.getAndClearAnnotationDragged()) {
+        if (pointAnnotations.getAndClearAnnotationDragged()) {
             return true
         }
         val screenPoint = mMap?.pixelForCoordinate(point)
@@ -1565,154 +1556,5 @@ fun OrnamentSettings.setPosAndMargins(posAndMargins: ReadableMap?) {
     this.margins = margins
 }
 
-class RNMBXPointAnnotationManager(val mapView: MapView) {
-    val manager: PointAnnotationManager;
-    var annotationClicked = false
-    var annotationDragged = false
-
-    var selected: RNMBXPointAnnotation? = null
-
-    val annotations: MutableMap<String, RNMBXPointAnnotation> = hashMapOf()
-
-    init {
-        manager = mapView.annotations.createPointAnnotationManager(AnnotationConfig(layerId = "RNMBX-mapview-annotations"))
-        manager.addClickListener(OnPointAnnotationClickListener { pointAnnotation ->
-            onAnnotationClick(pointAnnotation)
-            false
-        })
-    }
-
-    fun getAndClearAnnotationClicked(): Boolean {
-        if (annotationClicked) {
-            annotationClicked = false
-            return true
-        }
-        return false
-    }
-
-    fun getAndClearAnnotationDragged(): Boolean {
-        if (annotationDragged) {
-            annotationDragged = false
-            return true
-        }
-        return false
-    }
-
-    fun lookup(point: PointAnnotation): RNMBXPointAnnotation? {
-        for (annotation in annotations.values) {
-            if (point.id == annotation.mapboxID) {
-                return annotation;
-            }
-        }
-        Logger.e(LOG_TAG, "Failed to find RNMBXPointAnntoation for ${point.id}")
-        return null;
-    }
-
-    fun onAnnotationClick(pointAnnotation: RNMBXPointAnnotation) {
-        var oldSelected: RNMBXPointAnnotation? = selected
-        var newSelected: RNMBXPointAnnotation? = pointAnnotation
-
-        annotationClicked = true
-
-        if (newSelected == oldSelected) {
-            newSelected = null
-        }
-
-        manager.addDragListener(object : OnPointAnnotationDragListener {
-            override fun onAnnotationDragStarted(_annotation: Annotation<*>) {
-                annotationDragged = true;
-                var reactAnnotation: RNMBXPointAnnotation? = null
-                for (key in annotations.keys) {
-                    val annotation = annotations[key]
-                    val curMarkerID = annotation?.mapboxID
-                    if (_annotation.id == curMarkerID) {
-                        reactAnnotation = annotation
-                    }
-                }
-                reactAnnotation?.let { it.onDragStart() }
-            }
-
-            override fun onAnnotationDrag(_annotation: Annotation<*>) {
-                var reactAnnotation: RNMBXPointAnnotation? = null
-                for (key in annotations.keys) {
-                    val annotation = annotations[key]
-                    val curMarkerID = annotation?.mapboxID
-                    if (_annotation.id == curMarkerID) {
-                        reactAnnotation = annotation
-                    }
-                }
-                reactAnnotation?.let { it.onDrag() }
-            }
-
-            override fun onAnnotationDragFinished(_annotation: Annotation<*>) {
-                annotationDragged = false;
-                var reactAnnotation: RNMBXPointAnnotation? = null
-                for (key in annotations.keys) {
-                    val annotation = annotations[key]
-                    val curMarkerID = annotation?.mapboxID
-                    if (_annotation.id == curMarkerID) {
-                        reactAnnotation = annotation
-                    }
-                }
-                reactAnnotation?.let { it.onDragEnd() }
-            }
-        })
-
-        oldSelected?.let { deselectAnnotation(it) }
-        newSelected?.let { selectAnnotation(it) }
-
-    }
-
-    fun onAnnotationClick(point: PointAnnotation) {
-        lookup(point)?.let {
-            onAnnotationClick(it)
-        }
-    }
-
-    fun deselectSelectedAnnotation(): Boolean {
-        selected?.let {
-            deselectAnnotation(it)
-            return true
-        }
-        return false
-    }
-
-    fun selectAnnotation(annotation: RNMBXPointAnnotation) {
-        selected = annotation
-        annotation.doSelect(true)
-    }
-
-    fun deselectAnnotation(annotation: RNMBXPointAnnotation) {
-        selected = null
-        annotation.doDeselect()
-    }
-
-    fun remove(annotation: RNMBXPointAnnotation) {
-        if (annotation == selected) {
-            selected = null
-        }
-        annotations.remove(annotation.iD)
-    }
-
-    fun delete(annotation: PointAnnotation) {
-        manager.delete(annotation)
-    }
-
-    fun update(annotation: PointAnnotation) {
-        manager.update(annotation)
-    }
-
-    fun create(options: PointAnnotationOptions): PointAnnotation {
-        return manager.create(options)
-    }
-
-    fun add(annotation: RNMBXPointAnnotation) {
-        annotations[annotation.iD!!] = annotation
-    }
-
-    companion object {
-        const val LOG_TAG = "RNMBXPointAnnotationManager";
-    }
-}
